requirements.txt
flask
openai
pycryptodome

